+++ packages/dart/npt_flutter/lib/widgets/switch_atsign_button.dart
@@ -11,7 +11,10 @@ import 'package:npt_flutter/features/onboarding/util/post_onboard.dart';
 import 'package:npt_flutter/features/onboarding/util/pre_offboard.dart';
 import 'package:npt_flutter/features/profile_list/cubit/profiles_running_cubit.dart';
 import 'package:npt_flutter/features/profile_list/widgets/connected_profiles_dialog.dart';
+import 'package:npt_flutter/home_wrapper_widget.dart';
 import 'package:npt_flutter/localization/app_localizations.dart';
+import 'package:npt_flutter/pages/loading_page.dart';
+import 'package:npt_flutter/routes.dart';
 import 'package:npt_flutter/styles/app_color.dart';
 import 'package:npt_flutter/styles/sizes.dart';
 import 'package:npt_flutter/util/at_client_methods.dart';
@@ -47,15 +50,20 @@ class SwitchAtsignButton extends StatelessWidget {
               ),
               borderRadius: BorderRadius.circular(Sizes.p8),
             ),
-            items: (atSignList ?? [])
-                .map(
-                  (atSign) => PopupMenuItem<String>(
-                    padding: const EdgeInsets.all(Sizes.p0),
-                    value: atSign,
-                    child: _HoverableMenuItem(atSign: atSign),
-                  ),
-                )
-                .toList(),
+            items: [
+              ...(atSignList ?? []).map(
+                (atSign) => PopupMenuItem<String>(
+                  padding: const EdgeInsets.all(Sizes.p0),
+                  value: atSign,
+                  child: _HoverableMenuItem(atSign: atSign),
+                ),
+              ),
+              PopupMenuItem<String>(
+                padding: const EdgeInsets.all(Sizes.p0),
+                value: strings.signout,
+                child: _HoverableMenuItem(atSign: strings.signout),
+              ),
+            ],
           );
 
           if (selectedAtSign is! String) {
@@ -82,6 +90,24 @@ class SwitchAtsignButton extends StatelessWidget {
           if (context.mounted && isProfileConnected) {
             // If a proifle is connected, exit.
             return;
+          }
+
+          // If User selects signout, navigate to loading page and signout
+          // Otherwise, switch to the selected atsign
+          if (selectedAtSign == strings.signout) {
+            wrapperNav.currentState!.pushAndRemoveUntil(
+              MaterialPageRoute(builder: (context) => const LoadingPage()),
+              (route) => false,
+            );
+            await preSignout();
+            if (context.mounted) {
+              Navigator.of(
+                context,
+                rootNavigator: true,
+              ).pushNamedAndRemoveUntil(Routes.onboarding, (route) => false);
+
+              return;
+            }
           }
           final currentContext = App.navState.currentContext!;
 
@@ -152,6 +178,7 @@ class SwitchAtsignButton extends StatelessWidget {
 
 class _HoverableMenuItem extends StatefulWidget {
   final String atSign;
+
   const _HoverableMenuItem({required this.atSign});
   @override
   State<_HoverableMenuItem> createState() => _HoverableMenuItemState();
@@ -161,6 +188,7 @@ class _HoverableMenuItemState extends State<_HoverableMenuItem> {
   bool _hovering = false;
   @override
   Widget build(BuildContext context) {
+    final strings = AppLocalizations.of(context)!;
     return MouseRegion(
       onEnter: (_) => setState(() => _hovering = true),
       onExit: (_) => setState(() => _hovering = false),
@@ -179,9 +207,11 @@ class _HoverableMenuItemState extends State<_HoverableMenuItem> {
                   RichText(
                     text: TextSpan(
                       children: [
-                        const TextSpan(
-                          text: '@',
-                          style: TextStyle(color: AppColor.primaryColor),
+                        TextSpan(
+                          text: !widget.atSign.contains(strings.signout)
+                              ? '@'
+                              : '   ',
+                          style: const TextStyle(color: AppColor.primaryColor),
                         ),
                         TextSpan(
                           text: widget.atSign.split('@').last,